Iterative latent-class analysis

This function implements an iterative latent class analysis (ILCA; Jiang, 2019) approach to estimating attributes for cognitive diagnosis.

ILCA(dat, Q, seed.num = 5)

Arguments

dat
A required binary item response matrix.

Q
A required binary item and attribute association matrix.

seed.num
seed number

Value

Estimated attribute profiles.

References

Jiang, Z. (2019). Using the iterative latent-class analysis approach to improve attribute accuracy in diagnostic classification models. Behavior research methods, 1-10.

Examples

ILCA(sim10GDINA$simdat, sim10GDINA$simQ)

#>         A1 A2 A3
#>    [1,]  1  0  1
#>    [2,]  1  1  1
#>    [3,]  0  1  1
#>    [4,]  1  1  1
#>    [5,]  0  0  1
#>    [6,]  1  0  0
#>    [7,]  1  1  0
#>    [8,]  0  0  0
#>    [9,]  0  0  1
#>   [10,]  0  1  1
#>   [11,]  0  0  1
#>   [12,]  1  0  1
#>   [13,]  1  0  1
#>   [14,]  0  0  0
#>   [15,]  0  0  0
#>   [16,]  0  0  1
#>   [17,]  0  0  1
#>   [18,]  0  0  1
#>   [19,]  1  0  1
#>   [20,]  1  1  1
#>   [21,]  0  0  1
#>   [22,]  1  0  0
#>   [23,]  1  1  1
#>   [24,]  1  0  1
#>   [25,]  1  0  1
#>   [26,]  0  0  1
#>   [27,]  1  0  1
#>   [28,]  1  1  0
#>   [29,]  1  0  0
#>   [30,]  0  0  1
#>   [31,]  0  1  1
#>   [32,]  1  1  0
#>   [33,]  1  0  0
#>   [34,]  1  1  1
#>   [35,]  1  1  0
#>   [36,]  0  1  0
#>   [37,]  0  1  1
#>   [38,]  1  1  1
#>   [39,]  1  1  0
#>   [40,]  1  0  0
#>   [41,]  0  1  1
#>   [42,]  0  0  1
#>   [43,]  1  1  1
#>   [44,]  0  1  1
#>   [45,]  0  1  0
#>   [46,]  0  1  0
#>   [47,]  0  0  0
#>   [48,]  0  0  0
#>   [49,]  0  0  0
#>   [50,]  1  0  1
#>   [51,]  1  1  1
#>   [52,]  0  1  1
#>   [53,]  0  1  0
#>   [54,]  1  0  0
#>   [55,]  1  0  1
#>   [56,]  0  0  0
#>   [57,]  1  0  0
#>   [58,]  0  0  1
#>   [59,]  1  0  1
#>   [60,]  1  0  1
#>   [61,]  0  1  1
#>   [62,]  0  1  0
#>   [63,]  1  1  1
#>   [64,]  0  1  1
#>   [65,]  1  1  1
#>   [66,]  1  0  0
#>   [67,]  1  1  1
#>   [68,]  1  0  0
#>   [69,]  1  1  0
#>   [70,]  1  1  1
#>   [71,]  1  0  1
#>   [72,]  1  1  0
#>   [73,]  0  0  0
#>   [74,]  0  1  0
#>   [75,]  0  0  0
#>   [76,]  0  0  0
#>   [77,]  1  1  1
#>   [78,]  1  0  1
#>   [79,]  1  1  0
#>   [80,]  1  0  0
#>   [81,]  0  1  1
#>   [82,]  1  1  0
#>   [83,]  0  1  0
#>   [84,]  0  0  0
#>   [85,]  1  0  0
#>   [86,]  1  1  0
#>   [87,]  0  1  1
#>   [88,]  1  0  1
#>   [89,]  0  1  1
#>   [90,]  0  0  0
#>   [91,]  1  1  1
#>   [92,]  0  1  1
#>   [93,]  0  0  1
#>   [94,]  1  1  0
#>   [95,]  1  0  1
#>   [96,]  1  1  0
#>   [97,]  1  0  1
#>   [98,]  1  0  1
#>   [99,]  0  0  0
#>  [100,]  0  0  1
#>  [101,]  0  0  0
#>  [102,]  0  1  0
#>  [103,]  1  1  1
#>  [104,]  1  1  0
#>  [105,]  0  1  1
#>  [106,]  1  1  1
#>  [107,]  1  0  1
#>  [108,]  0  1  1
#>  [109,]  0  1  0
#>  [110,]  1  0  0
#>  [111,]  1  1  0
#>  [112,]  0  1  1
#>  [113,]  1  1  1
#>  [114,]  1  0  1
#>  [115,]  0  1  1
#>  [116,]  1  1  1
#>  [117,]  1  1  1
#>  [118,]  1  1  0
#>  [119,]  0  0  0
#>  [120,]  0  1  0
#>  [121,]  1  1  0
#>  [122,]  1  1  0
#>  [123,]  0  0  0
#>  [124,]  1  0  1
#>  [125,]  1  0  1
#>  [126,]  0  1  1
#>  [127,]  1  0  1
#>  [128,]  1  0  0
#>  [129,]  0  1  0
#>  [130,]  0  1  1
#>  [131,]  0  0  0
#>  [132,]  0  0  1
#>  [133,]  0  1  1
#>  [134,]  1  0  1
#>  [135,]  1  1  1
#>  [136,]  0  1  1
#>  [137,]  1  0  1
#>  [138,]  1  1  0
#>  [139,]  0  0  0
#>  [140,]  0  1  1
#>  [141,]  1  0  0
#>  [142,]  1  0  1
#>  [143,]  0  1  0
#>  [144,]  1  0  1
#>  [145,]  1  0  0
#>  [146,]  1  0  1
#>  [147,]  1  1  1
#>  [148,]  0  0  1
#>  [149,]  1  1  1
#>  [150,]  0  1  0
#>  [151,]  0  1  1
#>  [152,]  1  1  0
#>  [153,]  1  1  1
#>  [154,]  0  0  0
#>  [155,]  1  0  0
#>  [156,]  0  1  1
#>  [157,]  1  0  1
#>  [158,]  0  0  1
#>  [159,]  1  1  1
#>  [160,]  1  1  0
#>  [161,]  1  0  1
#>  [162,]  0  0  0
#>  [163,]  0  1  1
#>  [164,]  0  1  1
#>  [165,]  0  0  0
#>  [166,]  0  0  1
#>  [167,]  1  1  1
#>  [168,]  1  0  0
#>  [169,]  0  1  1
#>  [170,]  0  0  1
#>  [171,]  1  1  1
#>  [172,]  0  0  0
#>  [173,]  1  0  1
#>  [174,]  0  1  0
#>  [175,]  0  1  0
#>  [176,]  0  1  1
#>  [177,]  1  0  0
#>  [178,]  1  0  1
#>  [179,]  0  1  1
#>  [180,]  0  1  1
#>  [181,]  1  0  0
#>  [182,]  0  0  1
#>  [183,]  1  0  0
#>  [184,]  0  1  0
#>  [185,]  1  1  1
#>  [186,]  0  1  1
#>  [187,]  0  0  1
#>  [188,]  1  1  1
#>  [189,]  1  0  0
#>  [190,]  1  0  0
#>  [191,]  0  1  1
#>  [192,]  1  1  1
#>  [193,]  0  1  0
#>  [194,]  0  0  0
#>  [195,]  1  1  1
#>  [196,]  0  0  0
#>  [197,]  0  0  1
#>  [198,]  1  1  1
#>  [199,]  1  0  0
#>  [200,]  0  0  1
#>  [201,]  1  1  0
#>  [202,]  1  1  1
#>  [203,]  0  0  0
#>  [204,]  1  1  0
#>  [205,]  1  0  1
#>  [206,]  0  0  1
#>  [207,]  0  0  0
#>  [208,]  0  0  0
#>  [209,]  1  1  0
#>  [210,]  1  1  1
#>  [211,]  0  0  0
#>  [212,]  1  0  1
#>  [213,]  0  0  1
#>  [214,]  0  0  0
#>  [215,]  1  1  1
#>  [216,]  0  0  0
#>  [217,]  0  1  0
#>  [218,]  1  1  0
#>  [219,]  1  1  0
#>  [220,]  1  0  0
#>  [221,]  0  0  0
#>  [222,]  0  0  1
#>  [223,]  0  0  0
#>  [224,]  1  1  0
#>  [225,]  1  1  1
#>  [226,]  1  0  0
#>  [227,]  1  1  1
#>  [228,]  0  0  1
#>  [229,]  0  0  1
#>  [230,]  1  1  0
#>  [231,]  1  1  0
#>  [232,]  1  1  0
#>  [233,]  0  0  1
#>  [234,]  1  0  1
#>  [235,]  1  0  1
#>  [236,]  1  1  0
#>  [237,]  1  1  1
#>  [238,]  1  1  1
#>  [239,]  0  0  1
#>  [240,]  0  0  1
#>  [241,]  1  1  1
#>  [242,]  1  1  1
#>  [243,]  1  1  1
#>  [244,]  0  1  1
#>  [245,]  1  0  1
#>  [246,]  1  0  1
#>  [247,]  1  1  1
#>  [248,]  1  1  0
#>  [249,]  0  1  0
#>  [250,]  0  1  1
#>  [251,]  0  0  0
#>  [252,]  0  1  1
#>  [253,]  0  0  0
#>  [254,]  0  1  0
#>  [255,]  1  1  1
#>  [256,]  0  1  1
#>  [257,]  0  0  0
#>  [258,]  1  0  0
#>  [259,]  0  0  0
#>  [260,]  0  1  1
#>  [261,]  1  1  1
#>  [262,]  1  1  1
#>  [263,]  0  1  1
#>  [264,]  1  1  1
#>  [265,]  0  1  1
#>  [266,]  0  0  1
#>  [267,]  0  0  0
#>  [268,]  0  1  1
#>  [269,]  1  1  1
#>  [270,]  1  0  1
#>  [271,]  1  0  0
#>  [272,]  0  1  0
#>  [273,]  1  1  0
#>  [274,]  1  1  1
#>  [275,]  1  0  1
#>  [276,]  0  0  1
#>  [277,]  1  1  0
#>  [278,]  0  0  0
#>  [279,]  1  1  0
#>  [280,]  0  1  1
#>  [281,]  1  1  0
#>  [282,]  0  1  1
#>  [283,]  0  1  0
#>  [284,]  0  0  0
#>  [285,]  0  1  0
#>  [286,]  0  0  1
#>  [287,]  0  0  0
#>  [288,]  0  0  1
#>  [289,]  1  1  0
#>  [290,]  1  1  0
#>  [291,]  0  0  1
#>  [292,]  0  1  1
#>  [293,]  0  0  0
#>  [294,]  0  1  0
#>  [295,]  1  1  1
#>  [296,]  0  0  0
#>  [297,]  0  1  1
#>  [298,]  1  0  0
#>  [299,]  1  1  0
#>  [300,]  0  0  0
#>  [301,]  0  0  0
#>  [302,]  1  1  0
#>  [303,]  1  0  1
#>  [304,]  1  0  1
#>  [305,]  1  1  0
#>  [306,]  0  0  1
#>  [307,]  1  0  0
#>  [308,]  1  0  0
#>  [309,]  0  0  0
#>  [310,]  0  1  0
#>  [311,]  1  0  1
#>  [312,]  0  0  1
#>  [313,]  1  0  1
#>  [314,]  1  0  0
#>  [315,]  0  0  0
#>  [316,]  1  1  1
#>  [317,]  1  1  1
#>  [318,]  1  1  0
#>  [319,]  0  1  0
#>  [320,]  0  1  1
#>  [321,]  1  0  0
#>  [322,]  1  0  1
#>  [323,]  1  1  0
#>  [324,]  0  0  1
#>  [325,]  0  0  1
#>  [326,]  1  1  1
#>  [327,]  0  1  0
#>  [328,]  0  0  0
#>  [329,]  1  0  0
#>  [330,]  1  0  0
#>  [331,]  0  0  1
#>  [332,]  0  1  1
#>  [333,]  1  0  1
#>  [334,]  1  0  1
#>  [335,]  0  1  0
#>  [336,]  0  0  0
#>  [337,]  1  0  1
#>  [338,]  1  0  1
#>  [339,]  0  0  1
#>  [340,]  1  1  0
#>  [341,]  1  1  0
#>  [342,]  1  0  1
#>  [343,]  1  1  1
#>  [344,]  0  1  1
#>  [345,]  1  1  0
#>  [346,]  1  0  1
#>  [347,]  0  1  1
#>  [348,]  1  1  1
#>  [349,]  0  0  0
#>  [350,]  1  1  1
#>  [351,]  1  0  1
#>  [352,]  1  1  0
#>  [353,]  0  1  1
#>  [354,]  1  0  1
#>  [355,]  1  0  1
#>  [356,]  1  0  1
#>  [357,]  1  0  0
#>  [358,]  1  1  0
#>  [359,]  0  0  0
#>  [360,]  0  1  1
#>  [361,]  1  1  0
#>  [362,]  0  0  1
#>  [363,]  0  0  0
#>  [364,]  1  0  0
#>  [365,]  0  1  0
#>  [366,]  0  1  0
#>  [367,]  1  1  1
#>  [368,]  0  0  0
#>  [369,]  1  0  1
#>  [370,]  1  1  0
#>  [371,]  1  1  0
#>  [372,]  0  1  1
#>  [373,]  1  1  1
#>  [374,]  1  1  0
#>  [375,]  1  0  1
#>  [376,]  0  0  1
#>  [377,]  0  0  1
#>  [378,]  1  1  1
#>  [379,]  0  1  0
#>  [380,]  0  0  1
#>  [381,]  1  0  1
#>  [382,]  1  1  1
#>  [383,]  0  1  0
#>  [384,]  1  0  1
#>  [385,]  0  1  1
#>  [386,]  0  1  1
#>  [387,]  1  1  0
#>  [388,]  1  0  0
#>  [389,]  1  1  1
#>  [390,]  0  0  1
#>  [391,]  0  1  1
#>  [392,]  0  1  1
#>  [393,]  0  1  1
#>  [394,]  1  0  0
#>  [395,]  0  1  0
#>  [396,]  0  0  0
#>  [397,]  0  1  0
#>  [398,]  1  0  1
#>  [399,]  1  1  0
#>  [400,]  1  1  0
#>  [401,]  0  0  0
#>  [402,]  0  0  1
#>  [403,]  0  1  0
#>  [404,]  1  0  0
#>  [405,]  1  1  0
#>  [406,]  1  0  1
#>  [407,]  0  1  1
#>  [408,]  1  0  1
#>  [409,]  1  1  1
#>  [410,]  0  0  0
#>  [411,]  1  1  0
#>  [412,]  1  0  1
#>  [413,]  0  0  0
#>  [414,]  0  0  1
#>  [415,]  1  0  1
#>  [416,]  0  1  1
#>  [417,]  1  1  0
#>  [418,]  0  0  1
#>  [419,]  0  0  1
#>  [420,]  0  0  0
#>  [421,]  0  1  1
#>  [422,]  0  1  0
#>  [423,]  0  0  0
#>  [424,]  0  1  0
#>  [425,]  0  0  1
#>  [426,]  0  1  1
#>  [427,]  1  1  1
#>  [428,]  0  1  0
#>  [429,]  0  0  0
#>  [430,]  1  1  0
#>  [431,]  1  1  0
#>  [432,]  0  0  0
#>  [433,]  1  0  0
#>  [434,]  0  0  0
#>  [435,]  0  0  0
#>  [436,]  1  1  0
#>  [437,]  0  0  0
#>  [438,]  0  1  1
#>  [439,]  1  1  0
#>  [440,]  0  0  0
#>  [441,]  0  1  1
#>  [442,]  1  1  1
#>  [443,]  0  0  0
#>  [444,]  1  1  1
#>  [445,]  0  1  1
#>  [446,]  1  1  1
#>  [447,]  0  0  1
#>  [448,]  0  0  0
#>  [449,]  1  1  1
#>  [450,]  1  1  1
#>  [451,]  1  0  1
#>  [452,]  0  0  1
#>  [453,]  0  0  1
#>  [454,]  0  1  1
#>  [455,]  0  1  0
#>  [456,]  1  0  1
#>  [457,]  1  1  1
#>  [458,]  0  1  0
#>  [459,]  1  1  0
#>  [460,]  1  0  0
#>  [461,]  1  1  1
#>  [462,]  0  1  1
#>  [463,]  1  0  0
#>  [464,]  0  1  1
#>  [465,]  1  1  0
#>  [466,]  0  0  1
#>  [467,]  1  1  0
#>  [468,]  0  0  1
#>  [469,]  0  1  1
#>  [470,]  0  0  1
#>  [471,]  0  1  0
#>  [472,]  0  0  1
#>  [473,]  1  1  1
#>  [474,]  1  0  1
#>  [475,]  0  1  0
#>  [476,]  0  1  0
#>  [477,]  1  0  1
#>  [478,]  0  1  1
#>  [479,]  1  0  0
#>  [480,]  0  0  0
#>  [481,]  1  0  0
#>  [482,]  0  1  0
#>  [483,]  1  1  0
#>  [484,]  1  1  1
#>  [485,]  0  0  1
#>  [486,]  1  0  1
#>  [487,]  1  0  0
#>  [488,]  0  1  1
#>  [489,]  0  0  1
#>  [490,]  1  1  1
#>  [491,]  0  1  1
#>  [492,]  1  1  0
#>  [493,]  1  1  1
#>  [494,]  1  1  0
#>  [495,]  0  0  1
#>  [496,]  0  1  0
#>  [497,]  1  0  1
#>  [498,]  0  1  1
#>  [499,]  0  1  1
#>  [500,]  0  1  1
#>  [501,]  1  0  0
#>  [502,]  0  1  1
#>  [503,]  0  1  1
#>  [504,]  1  1  1
#>  [505,]  1  0  1
#>  [506,]  0  1  1
#>  [507,]  0  0  0
#>  [508,]  1  0  1
#>  [509,]  0  0  1
#>  [510,]  0  0  1
#>  [511,]  1  0  1
#>  [512,]  0  1  0
#>  [513,]  0  1  0
#>  [514,]  1  0  1
#>  [515,]  0  1  0
#>  [516,]  1  0  1
#>  [517,]  0  1  0
#>  [518,]  1  0  0
#>  [519,]  1  0  0
#>  [520,]  0  1  1
#>  [521,]  1  1  1
#>  [522,]  1  0  0
#>  [523,]  1  1  1
#>  [524,]  1  1  0
#>  [525,]  0  0  1
#>  [526,]  0  1  0
#>  [527,]  0  1  0
#>  [528,]  1  1  0
#>  [529,]  1  1  1
#>  [530,]  0  1  1
#>  [531,]  0  0  1
#>  [532,]  0  1  0
#>  [533,]  1  1  0
#>  [534,]  1  0  1
#>  [535,]  1  0  1
#>  [536,]  1  1  0
#>  [537,]  0  0  1
#>  [538,]  1  1  1
#>  [539,]  0  1  0
#>  [540,]  0  0  1
#>  [541,]  0  1  0
#>  [542,]  0  1  0
#>  [543,]  1  0  0
#>  [544,]  1  1  0
#>  [545,]  0  0  1
#>  [546,]  1  1  0
#>  [547,]  0  0  0
#>  [548,]  1  0  0
#>  [549,]  1  1  0
#>  [550,]  0  1  1
#>  [551,]  0  1  1
#>  [552,]  0  0  0
#>  [553,]  1  1  0
#>  [554,]  0  1  1
#>  [555,]  0  0  1
#>  [556,]  0  0  0
#>  [557,]  0  1  0
#>  [558,]  0  0  0
#>  [559,]  1  0  1
#>  [560,]  0  1  0
#>  [561,]  0  0  1
#>  [562,]  0  0  0
#>  [563,]  0  1  0
#>  [564,]  1  0  1
#>  [565,]  0  0  1
#>  [566,]  0  1  0
#>  [567,]  1  1  1
#>  [568,]  0  1  0
#>  [569,]  1  0  0
#>  [570,]  1  0  0
#>  [571,]  1  0  0
#>  [572,]  1  0  1
#>  [573,]  1  1  0
#>  [574,]  0  0  0
#>  [575,]  1  0  0
#>  [576,]  0  0  1
#>  [577,]  0  1  1
#>  [578,]  1  0  0
#>  [579,]  0  1  1
#>  [580,]  1  1  0
#>  [581,]  0  1  0
#>  [582,]  0  1  1
#>  [583,]  1  0  0
#>  [584,]  0  1  0
#>  [585,]  0  0  1
#>  [586,]  1  1  1
#>  [587,]  1  0  1
#>  [588,]  1  0  1
#>  [589,]  0  0  0
#>  [590,]  0  0  1
#>  [591,]  1  0  0
#>  [592,]  0  0  1
#>  [593,]  0  0  0
#>  [594,]  1  0  1
#>  [595,]  0  1  0
#>  [596,]  1  0  1
#>  [597,]  1  0  1
#>  [598,]  1  1  1
#>  [599,]  0  1  1
#>  [600,]  0  1  1
#>  [601,]  1  0  1
#>  [602,]  0  0  0
#>  [603,]  1  1  0
#>  [604,]  1  1  0
#>  [605,]  0  1  0
#>  [606,]  0  1  1
#>  [607,]  1  1  1
#>  [608,]  0  1  1
#>  [609,]  0  0  0
#>  [610,]  0  0  1
#>  [611,]  1  1  0
#>  [612,]  1  0  0
#>  [613,]  1  0  1
#>  [614,]  1  0  0
#>  [615,]  1  0  1
#>  [616,]  0  0  0
#>  [617,]  0  0  0
#>  [618,]  1  1  0
#>  [619,]  1  0  1
#>  [620,]  1  1  0
#>  [621,]  1  0  0
#>  [622,]  0  1  1
#>  [623,]  1  0  1
#>  [624,]  1  0  0
#>  [625,]  1  1  1
#>  [626,]  1  1  1
#>  [627,]  1  1  1
#>  [628,]  1  0  1
#>  [629,]  1  0  1
#>  [630,]  0  1  0
#>  [631,]  0  0  1
#>  [632,]  0  0  1
#>  [633,]  0  1  0
#>  [634,]  0  0  1
#>  [635,]  0  0  0
#>  [636,]  1  0  1
#>  [637,]  1  0  1
#>  [638,]  0  0  0
#>  [639,]  1  0  0
#>  [640,]  1  0  0
#>  [641,]  1  0  0
#>  [642,]  1  1  0
#>  [643,]  1  0  1
#>  [644,]  1  1  0
#>  [645,]  1  1  1
#>  [646,]  1  1  0
#>  [647,]  0  1  1
#>  [648,]  1  1  1
#>  [649,]  0  0  0
#>  [650,]  0  1  0
#>  [651,]  0  1  1
#>  [652,]  0  1  0
#>  [653,]  1  0  1
#>  [654,]  1  1  0
#>  [655,]  1  0  1
#>  [656,]  1  1  1
#>  [657,]  1  0  0
#>  [658,]  0  1  0
#>  [659,]  1  0  1
#>  [660,]  1  0  0
#>  [661,]  1  1  1
#>  [662,]  0  0  0
#>  [663,]  1  0  1
#>  [664,]  1  1  1
#>  [665,]  1  0  1
#>  [666,]  0  0  1
#>  [667,]  0  1  1
#>  [668,]  0  1  0
#>  [669,]  1  0  1
#>  [670,]  1  1  0
#>  [671,]  1  0  1
#>  [672,]  0  0  0
#>  [673,]  0  0  0
#>  [674,]  1  1  0
#>  [675,]  1  1  1
#>  [676,]  1  0  0
#>  [677,]  1  0  1
#>  [678,]  0  1  0
#>  [679,]  1  0  1
#>  [680,]  0  0  0
#>  [681,]  1  1  1
#>  [682,]  1  0  0
#>  [683,]  0  0  1
#>  [684,]  0  1  1
#>  [685,]  0  1  1
#>  [686,]  1  1  0
#>  [687,]  1  0  0
#>  [688,]  1  1  1
#>  [689,]  0  0  1
#>  [690,]  0  0  1
#>  [691,]  0  0  1
#>  [692,]  1  1  1
#>  [693,]  0  0  0
#>  [694,]  0  0  0
#>  [695,]  1  0  1
#>  [696,]  1  1  1
#>  [697,]  0  0  0
#>  [698,]  0  1  1
#>  [699,]  0  1  1
#>  [700,]  0  0  1
#>  [701,]  1  1  0
#>  [702,]  0  1  0
#>  [703,]  0  0  1
#>  [704,]  0  1  0
#>  [705,]  0  1  0
#>  [706,]  0  0  1
#>  [707,]  1  0  0
#>  [708,]  0  1  1
#>  [709,]  0  1  1
#>  [710,]  0  1  0
#>  [711,]  1  1  0
#>  [712,]  0  1  1
#>  [713,]  1  0  1
#>  [714,]  0  1  0
#>  [715,]  1  1  0
#>  [716,]  0  0  1
#>  [717,]  1  0  1
#>  [718,]  1  1  1
#>  [719,]  1  0  0
#>  [720,]  0  1  0
#>  [721,]  1  1  0
#>  [722,]  0  0  0
#>  [723,]  0  0  0
#>  [724,]  0  0  0
#>  [725,]  1  1  0
#>  [726,]  0  0  1
#>  [727,]  0  0  1
#>  [728,]  0  0  0
#>  [729,]  1  0  0
#>  [730,]  1  0  0
#>  [731,]  0  1  1
#>  [732,]  1  0  1
#>  [733,]  0  1  0
#>  [734,]  0  1  1
#>  [735,]  1  1  1
#>  [736,]  0  1  0
#>  [737,]  1  0  0
#>  [738,]  0  0  0
#>  [739,]  0  0  1
#>  [740,]  1  0  1
#>  [741,]  0  0  0
#>  [742,]  0  0  0
#>  [743,]  1  1  1
#>  [744,]  1  1  1
#>  [745,]  1  0  1
#>  [746,]  0  0  0
#>  [747,]  1  0  1
#>  [748,]  0  1  0
#>  [749,]  0  0  1
#>  [750,]  1  1  1
#>  [751,]  1  0  0
#>  [752,]  1  1  0
#>  [753,]  1  1  1
#>  [754,]  0  0  1
#>  [755,]  0  0  0
#>  [756,]  1  0  0
#>  [757,]  0  1  1
#>  [758,]  1  0  1
#>  [759,]  1  1  1
#>  [760,]  0  0  0
#>  [761,]  0  1  0
#>  [762,]  0  1  0
#>  [763,]  0  0  1
#>  [764,]  1  1  0
#>  [765,]  1  0  1
#>  [766,]  1  1  1
#>  [767,]  0  1  0
#>  [768,]  1  0  1
#>  [769,]  1  0  1
#>  [770,]  1  0  0
#>  [771,]  0  0  1
#>  [772,]  0  1  1
#>  [773,]  0  1  1
#>  [774,]  0  0  0
#>  [775,]  1  1  1
#>  [776,]  0  0  1
#>  [777,]  0  0  0
#>  [778,]  0  0  0
#>  [779,]  1  1  0
#>  [780,]  0  1  1
#>  [781,]  0  1  1
#>  [782,]  0  0  1
#>  [783,]  1  0  1
#>  [784,]  0  0  0
#>  [785,]  1  1  0
#>  [786,]  1  1  1
#>  [787,]  0  1  0
#>  [788,]  0  1  0
#>  [789,]  1  1  0
#>  [790,]  0  1  0
#>  [791,]  1  1  0
#>  [792,]  0  0  0
#>  [793,]  1  0  1
#>  [794,]  0  0  1
#>  [795,]  1  1  0
#>  [796,]  0  1  1
#>  [797,]  0  0  1
#>  [798,]  0  0  1
#>  [799,]  0  0  0
#>  [800,]  0  0  1
#>  [801,]  1  0  1
#>  [802,]  1  1  1
#>  [803,]  1  0  0
#>  [804,]  0  0  0
#>  [805,]  1  0  0
#>  [806,]  0  0  0
#>  [807,]  1  0  0
#>  [808,]  1  0  0
#>  [809,]  0  0  0
#>  [810,]  0  1  0
#>  [811,]  0  1  1
#>  [812,]  0  0  1
#>  [813,]  0  0  1
#>  [814,]  1  0  0
#>  [815,]  1  1  0
#>  [816,]  0  1  0
#>  [817,]  0  0  0
#>  [818,]  1  1  1
#>  [819,]  1  0  0
#>  [820,]  1  1  0
#>  [821,]  1  0  0
#>  [822,]  1  0  0
#>  [823,]  0  0  1
#>  [824,]  1  0  1
#>  [825,]  1  1  1
#>  [826,]  1  1  0
#>  [827,]  1  1  0
#>  [828,]  1  0  1
#>  [829,]  1  0  0
#>  [830,]  0  0  0
#>  [831,]  0  0  1
#>  [832,]  1  0  1
#>  [833,]  1  0  1
#>  [834,]  1  0  1
#>  [835,]  1  0  0
#>  [836,]  0  0  1
#>  [837,]  1  0  0
#>  [838,]  1  0  1
#>  [839,]  0  1  1
#>  [840,]  1  0  0
#>  [841,]  1  0  1
#>  [842,]  0  0  0
#>  [843,]  1  0  0
#>  [844,]  1  1  0
#>  [845,]  0  1  0
#>  [846,]  1  1  0
#>  [847,]  1  1  0
#>  [848,]  0  0  0
#>  [849,]  1  1  1
#>  [850,]  0  0  1
#>  [851,]  0  1  0
#>  [852,]  1  1  1
#>  [853,]  1  1  1
#>  [854,]  0  1  0
#>  [855,]  1  0  0
#>  [856,]  0  1  0
#>  [857,]  0  1  1
#>  [858,]  1  0  0
#>  [859,]  1  0  1
#>  [860,]  1  1  0
#>  [861,]  1  0  0
#>  [862,]  0  0  0
#>  [863,]  0  0  1
#>  [864,]  0  1  1
#>  [865,]  1  1  0
#>  [866,]  1  0  1
#>  [867,]  0  1  0
#>  [868,]  0  0  0
#>  [869,]  1  0  1
#>  [870,]  1  1  1
#>  [871,]  1  1  0
#>  [872,]  1  0  0
#>  [873,]  1  1  1
#>  [874,]  1  1  0
#>  [875,]  1  1  1
#>  [876,]  0  0  1
#>  [877,]  0  1  1
#>  [878,]  0  0  0
#>  [879,]  1  1  1
#>  [880,]  1  1  0
#>  [881,]  1  0  1
#>  [882,]  0  0  0
#>  [883,]  1  0  1
#>  [884,]  0  0  1
#>  [885,]  1  1  0
#>  [886,]  1  0  1
#>  [887,]  1  0  1
#>  [888,]  0  1  0
#>  [889,]  1  1  1
#>  [890,]  0  0  0
#>  [891,]  1  1  1
#>  [892,]  1  0  0
#>  [893,]  1  0  0
#>  [894,]  0  0  0
#>  [895,]  0  0  0
#>  [896,]  1  0  0
#>  [897,]  1  0  1
#>  [898,]  0  1  0
#>  [899,]  1  0  0
#>  [900,]  0  1  0
#>  [901,]  1  1  1
#>  [902,]  0  0  0
#>  [903,]  0  1  1
#>  [904,]  0  1  0
#>  [905,]  0  1  1
#>  [906,]  0  0  0
#>  [907,]  0  1  0
#>  [908,]  1  1  1
#>  [909,]  0  1  1
#>  [910,]  1  0  0
#>  [911,]  1  1  1
#>  [912,]  1  0  1
#>  [913,]  1  1  1
#>  [914,]  1  0  1
#>  [915,]  0  0  0
#>  [916,]  0  1  0
#>  [917,]  1  0  0
#>  [918,]  0  1  0
#>  [919,]  1  0  1
#>  [920,]  0  1  0
#>  [921,]  0  0  0
#>  [922,]  1  0  1
#>  [923,]  0  0  1
#>  [924,]  1  1  0
#>  [925,]  1  1  0
#>  [926,]  0  1  0
#>  [927,]  0  0  0
#>  [928,]  0  0  0
#>  [929,]  0  0  1
#>  [930,]  0  1  0
#>  [931,]  1  0  0
#>  [932,]  1  0  1
#>  [933,]  1  0  0
#>  [934,]  1  1  1
#>  [935,]  1  0  0
#>  [936,]  1  1  0
#>  [937,]  0  0  0
#>  [938,]  0  1  0
#>  [939,]  0  0  1
#>  [940,]  1  1  0
#>  [941,]  0  1  1
#>  [942,]  1  1  0
#>  [943,]  0  0  1
#>  [944,]  0  1  1
#>  [945,]  0  0  0
#>  [946,]  1  1  1
#>  [947,]  1  0  1
#>  [948,]  1  1  0
#>  [949,]  0  1  1
#>  [950,]  1  1  0
#>  [951,]  0  1  0
#>  [952,]  0  1  0
#>  [953,]  0  1  0
#>  [954,]  1  0  0
#>  [955,]  0  1  0
#>  [956,]  0  0  1
#>  [957,]  0  0  1
#>  [958,]  1  0  1
#>  [959,]  0  0  1
#>  [960,]  0  1  0
#>  [961,]  0  0  1
#>  [962,]  1  1  0
#>  [963,]  0  0  1
#>  [964,]  1  0  1
#>  [965,]  0  0  1
#>  [966,]  1  1  0
#>  [967,]  1  0  1
#>  [968,]  1  0  1
#>  [969,]  0  0  0
#>  [970,]  0  1  1
#>  [971,]  1  1  0
#>  [972,]  1  1  1
#>  [973,]  0  1  1
#>  [974,]  1  1  1
#>  [975,]  1  1  0
#>  [976,]  1  0  1
#>  [977,]  1  1  0
#>  [978,]  0  0  1
#>  [979,]  1  1  0
#>  [980,]  1  1  1
#>  [981,]  0  0  0
#>  [982,]  0  0  1
#>  [983,]  0  1  1
#>  [984,]  0  1  1
#>  [985,]  0  0  1
#>  [986,]  0  0  1
#>  [987,]  0  0  0
#>  [988,]  1  0  1
#>  [989,]  1  0  1
#>  [990,]  0  0  0
#>  [991,]  0  1  0
#>  [992,]  0  1  0
#>  [993,]  0  0  0
#>  [994,]  1  1  0
#>  [995,]  1  0  0
#>  [996,]  0  1  1
#>  [997,]  0  0  1
#>  [998,]  1  1  0
#>  [999,]  0  1  1
#> [1000,]  1  0  0