Generate unrestricted Qc matrix from an restricted Qc matrix

Generate unrestricted Qc matrix from an restricted Qc matrix

unrestrQ(Qc)

Arguments

Qc

an restricted Qc matrix

Value

an unrestricted Qc matrix

Examples

Qc <- sim21seqDINA$simQc
Qc
#>    Item Cat A1 A2 A3 A4 A5
#> 1     1   1  1  0  0  0  0
#> 2     1   2  0  1  0  0  0
#> 3     2   1  0  0  1  0  0
#> 4     2   2  0  0  0  1  0
#> 5     3   1  0  0  0  0  1
#> 6     3   2  1  0  0  0  0
#> 7     4   1  0  0  0  0  1
#> 8     4   2  0  0  0  1  0
#> 9     5   1  0  0  1  0  0
#> 10    5   2  0  1  0  0  0
#> 11    6   1  1  0  0  0  0
#> 12    6   2  0  1  1  0  0
#> 13    7   1  0  0  1  0  0
#> 14    7   2  0  0  0  1  1
#> 15    8   1  0  0  0  0  1
#> 16    8   2  1  1  0  0  0
#> 17    9   1  0  0  0  1  1
#> 18    9   2  0  0  1  0  0
#> 19   10   1  0  1  0  1  0
#> 20   10   2  1  0  0  0  0
#> 21   11   1  1  1  0  0  0
#> 22   11   2  0  0  0  0  1
#> 23   12   1  1  1  1  0  0
#> 24   12   2  0  0  0  1  1
#> 25   13   1  1  1  0  0  0
#> 26   13   2  0  0  1  1  1
#> 27   14   1  1  0  1  0  0
#> 28   14   2  0  0  0  1  0
#> 29   14   3  0  0  0  0  1
#> 30   15   1  0  0  0  0  1
#> 31   15   2  0  0  1  1  0
#> 32   15   3  0  1  0  0  0
#> 33   16   1  1  0  0  0  0
#> 34   16   2  0  1  0  0  0
#> 35   16   3  0  0  1  1  0
#> 36   17   1  1  0  0  0  0
#> 37   18   1  0  1  0  0  0
#> 38   19   1  0  0  1  0  0
#> 39   20   1  0  0  0  1  0
#> 40   21   1  0  0  0  0  1
unrestrQ(Qc)
#>    Item Cat A1 A2 A3 A4 A5
#> 1     1   1  1  1  0  0  0
#> 2     1   2  1  1  0  0  0
#> 3     2   1  0  0  1  1  0
#> 4     2   2  0  0  1  1  0
#> 5     3   1  1  0  0  0  1
#> 6     3   2  1  0  0  0  1
#> 7     4   1  0  0  0  1  1
#> 8     4   2  0  0  0  1  1
#> 9     5   1  0  1  1  0  0
#> 10    5   2  0  1  1  0  0
#> 11    6   1  1  1  1  0  0
#> 12    6   2  1  1  1  0  0
#> 13    7   1  0  0  1  1  1
#> 14    7   2  0  0  1  1  1
#> 15    8   1  1  1  0  0  1
#> 16    8   2  1  1  0  0  1
#> 17    9   1  0  0  1  1  1
#> 18    9   2  0  0  1  1  1
#> 19   10   1  1  1  0  1  0
#> 20   10   2  1  1  0  1  0
#> 21   11   1  1  1  0  0  1
#> 22   11   2  1  1  0  0  1
#> 23   12   1  1  1  1  1  1
#> 24   12   2  1  1  1  1  1
#> 25   13   1  1  1  1  1  1
#> 26   13   2  1  1  1  1  1
#> 27   14   1  1  0  1  1  1
#> 28   14   2  1  0  1  1  1
#> 29   14   3  1  0  1  1  1
#> 30   15   1  0  1  1  1  1
#> 31   15   2  0  1  1  1  1
#> 32   15   3  0  1  1  1  1
#> 33   16   1  1  1  1  1  0
#> 34   16   2  1  1  1  1  0
#> 35   16   3  1  1  1  1  0
#> 36   17   1  1  0  0  0  0
#> 37   18   1  0  1  0  0  0
#> 38   19   1  0  0  1  0  0
#> 39   20   1  0  0  0  1  0
#> 40   21   1  0  0  0  0  1