Q-matrix validation

Q-matrix validation for the G-DINA model based on PVAF (de la Torre & Chiu, 2016) or stepwise Wald test (Ma, 2017).

Qval(GDINA.obj, method = "PVAF", eps = 0.95, digits = 4,
  wald.args = list())

# S3 method for Qval
extract(object, what = c("sug.Q", "varsigma", "PVAF",
  "eps", "Q"), ...)

# S3 method for Qval
summary(object, ...)

Arguments

GDINA.obj
An estimated model object of class GDINA

method
which Q-matrix validation method is used?

eps
cutoff value for PVAF. 0.95 is the default.

digits
How many decimal places in each number? The default is 4.

wald.args
a list of arguments for the stepwise Wald test method.

SE.type

Type of covariance matrix for the Wald test

alpha.level

alpha level for the wald test

GDI

It can be 0, 1 or 2; 0 means GDI is not used to choose the attribute - when more than one attributes are significant, the one with the largest p-value will be selected; GDI=1 means the attribute with the largest GDI will be selected; GDI=2 means the q-vector with the largest GDI will be selected.

verbose

Print detailed information or not?

stepwise

TRUE for stepwise approach and FALSE for forward approach

object
Qval objects for S3 methods

what
argument for S3 method extract indicating what to extract; It can be "sug.Q" for suggested Q-matrix, "Q" for original Q-matrix, "varsigma" for varsigma index, and "PVAF" for PVAF.

...
additional arguments

Value

An object of class Qval. Elements that can be extracted using extract method include:

sug.Q

suggested Q-matrix

Q

original Q-matrix

varsigma

varsigma index

PVAF

PVAF

Methods (by generic)

extract: extract various elements from Qval objects

summary: print summary information

References

de la Torre, J. & Chiu, C-Y. (2016). A General Method of Empirical Q-matrix Validation. Psychometrika, 81, 253-273.

Ma, W. (2017). A Sequential Cognitive Diagnosis Model for Graded Response: Model Development, Q-matrix Validation, and Model Comparison. Unpublished doctoral dissertation. Rutgers, The State University of New Jersey.

See also

Examples

# NOT RUN {
dat <- sim10GDINA$simdat
Q <- sim10GDINA$simQ
Q[10,] <- c(0,1,0)
mod1 <- GDINA(dat = dat, Q = Q, model = "GDINA")
out <- Qval(mod1,eps = 0.95)
out
#If many entries are modified, you may want to check
#the PVAF plot using the function plotPVAF or
#to change eps. eps = 0.9 or 0.8 seems another two
#reasonable choices.
extract(out,what = "PVAF")
#See also:
extract(out,what = "varsigma")
extract(out,what = "sug.Q")

# Draw a mesa plot
plot(out,item=10,type="best",no.qvector=5)
# }