Calculating standard errors and variance-covariance matrix using bootstrap methods

This function conducts nonparametric and parametric bootstrap to calculate standard errors of model parameters. Parametric bootstrap is only applicable to single group models.

bootSE(GDINA.obj, bootsample = 50, type = "nonparametric",
  randomseed = 12345)

Arguments

GDINA.obj
an object of class GDINA

bootsample
the number of bootstrap samples

type
type of bootstrap method. Can be parametric or nonparametric

randomseed
random seed for resampling

Value

itemparm.se standard errors for item probability of success in list format

delta.se standard errors for delta parameters in list format

lambda.se standard errors for structural parameters of joint attribute distribution

boot.est resample estimates

Examples

# NOT RUN {
# For illustration, only 5 resamples are run
# results are definitely not reliable

dat <- sim30GDINA$simdat
Q <- sim30GDINA$simQ
fit <- GDINA(dat = dat, Q = Q, model = "GDINA",att.dist = "higher.order")
boot.fit <- bootSE(fit,bootsample = 5,randomseed=123)
boot.fit$delta.se
boot.fit$lambda.se
# }