extract elements from objects of various classes

A generic function to extract elements from objects of class GDINA, itemfit, modelcomp, Qval or simGDINA. This page gives the elements that can be extracted from the class GDINA. To see what can be extracted from itemfit, modelcomp, and Qval, go to the corresponding function help page.

Objects which can be extracted from GDINA objects include:

att.prior

attribute prior weights for calculating marginalized likelihood in the last EM iteration

discrim

GDINA discrimination index

designmatrix

A list of design matrices for each item/category

expectedCorrect

expected # of examinees in each latent group answering item correctly

expectedTotal

expected # of examinees in each latent group

higher.order

higher-order model specifications

linkfunc

link functions for each item

initial.catprob

initial item category probability parameters

prevalence

prevalence of each attribute

posterior.prob

posterior weights for each latent class

extract(object, what, ...)

Arguments

object
objects from class GDINA,itemfit, modelcomp, Qval or simGDINA

what
what to extract

...
additional arguments

Examples

# NOT RUN {
dat <- sim10GDINA$simdat
Q <- sim10GDINA$simQ
fit <- GDINA(dat = dat, Q = Q, model = "GDINA")
extract(fit,"discrim")
extract(fit,"designmatrix")
# }